+++ lib/l10n/extensions.dart
@@ -1,7 +1,5 @@
-import "package:flow/data/money.dart";
 import "package:flow/l10n/flow_localizations.dart";
 import "package:flutter/widgets.dart";
-import "package:intl/intl.dart";
 
 extension L10nHelper on BuildContext {
   FlowLocalizations get l => FlowLocalizations.of(this);
@@ -43,65 +41,4 @@ extension L10nStringHelper on String {
   /// Same as calling context.l.get([this])
   String tr([dynamic replace]) =>
       FlowLocalizations.getTransalation(this, replace: replace);
-}
-
-extension MoneyFormatters on Money {
-  String formatMoney({
-    bool includeCurrency = true,
-    bool useCurrencySymbol = true,
-    bool compact = false,
-    bool takeAbsoluteValue = false,
-    int? decimalDigits,
-  }) {
-    final num amountToFormat = takeAbsoluteValue ? amount.abs() : amount;
-    final String currencyToFormat = !includeCurrency ? "" : currency;
-    useCurrencySymbol = useCurrencySymbol && includeCurrency;
-
-    final String? symbol = useCurrencySymbol
-        ? NumberFormat.simpleCurrency(
-            locale: Intl.defaultLocale,
-            name: currencyToFormat,
-          ).currencySymbol
-        : null;
-
-    if (compact) {
-      return NumberFormat.compactCurrency(
-        locale: Intl.defaultLocale,
-        name: currencyToFormat,
-        symbol: symbol,
-        decimalDigits: decimalDigits,
-      ).format(amountToFormat);
-    }
-    return NumberFormat.currency(
-      locale: Intl.defaultLocale,
-      name: currencyToFormat,
-      symbol: symbol,
-      decimalDigits: decimalDigits,
-    ).format(amountToFormat);
-  }
-
-  /// Returns money-formatted string in the primary currency
-  /// in the default locale
-  ///
-  /// e.g., $420.69
-  String get money => formatMoney();
-
-  /// Returns compact money-formatted string in the primary
-  /// currency in the default locale
-  ///
-  /// e.g., $1.2M
-  String get moneyCompact => formatMoney(compact: true);
-
-  /// Returns money-formatted string (in the default locale)
-  ///
-  /// e.g., 467,000
-  String get moneyNoMarker => formatMoney(includeCurrency: false);
-
-  /// Returns money-formatted string (in the default locale)
-  ///
-  /// e.g., 1.2M
-  String get moneyNoMarkerCompact => formatMoney(
-        includeCurrency: false,
-        compact: true,
-      );
 }