+++ lib/widgets/general/markdown_view.dart
@@ -1,90 +1,71 @@
 import "dart:developer";
 
-import "package:flow/entity/transaction.dart";
 import "package:flow/l10n/extensions.dart";
-import "package:flow/routes/new_transaction/section.dart";
-import "package:flow/routes/utils/edit_markdown_page.dart";
 import "package:flow/theme/theme.dart";
 import "package:flow/utils/extensions/toast.dart";
 import "package:flow/utils/open_url.dart";
 import "package:flow/widgets/general/frame.dart";
-import "package:flow/widgets/general/markdown_view.dart";
 import "package:flutter/material.dart";
 import "package:flutter_markdown/flutter_markdown.dart";
-import "package:go_router/go_router.dart";
-import "package:material_symbols_icons/symbols.dart";
-import "package:simple_icons/simple_icons.dart";
 
-class DescriptionSection extends StatelessWidget {
+class MarkdownView extends StatelessWidget {
   final TextEditingController controller;
   final FocusNode? focusNode;
 
   final Function(String)? onChanged;
 
-  const DescriptionSection({
+  final bool allowTogglingCheckboxes;
+
+  const MarkdownView({
     super.key,
     required this.controller,
     this.focusNode,
     this.onChanged,
+    this.allowTogglingCheckboxes = false,
   });
 
   @override
   Widget build(BuildContext context) {
-    final bool noContent = controller.text.trim().isEmpty;
+    int checkboxCounter = 0;
 
-    return Section(
-      titleOverride: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Text("transaction.description".t(context)),
-          const SizedBox(width: 4.0),
-          Tooltip(
-            message: "transaction.description.markdownSupported".t(context),
-            child: Icon(
-              SimpleIcons.markdown,
-              size: 16.0,
-              fill: 0,
-              color: context.flowColors.semi,
-            ),
-          )
-        ],
+    return Frame(
+      child: Container(
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(16.0),
+          color: context.colorScheme.secondary,
+        ),
+        child: Markdown(
+          data: controller.text,
+          selectable: false,
+          shrinkWrap: true,
+          styleSheet: getStyleSheet(context),
+          checkboxBuilder: (value) {
+            final int index = checkboxCounter++;
+
+            return Checkbox.adaptive(
+              value: value,
+              onChanged: (newValue) => {
+                tryFlipCheckbox(index, newValue ?? !value),
+              },
+            );
+          },
+          onTapLink: (text, href, title) => onTapLink(
+            context,
+            text,
+            href,
+            title,
+          ),
+        ),
       ),
-      child: noContent
-          ? Align(
-              alignment: Alignment.topLeft,
-              child: Frame(
-                child: TextButton(
-                  onPressed: () => showEditModal(context),
-                  child: Text(
-                    "transaction.description.add".t(context),
-                  ),
-                ),
-              ),
-            )
-          : Stack(
-              children: [
-                MarkdownView(
-                  controller: controller,
-                  onChanged: onChanged,
-                  focusNode: focusNode,
-                  allowTogglingCheckboxes: true,
-                ),
-                Positioned(
-                  right: 24.0,
-                  top: 8.0,
-                  child: IconButton(
-                    isSelected: true,
-                    icon: Icon(Symbols.edit_rounded),
-                    onPressed: () => showEditModal(context),
-                    tooltip: "general.edit".t(context),
-                  ),
-                ),
-              ],
-            ),
     );
   }
 
   void tryFlipCheckbox(int index, bool value) {
+    if (!allowTogglingCheckboxes) {
+      log("[Flow] Cannot flip checkbox when toggling is disabled");
+      return;
+    }
+
     if (controller.text.contains("```")) {
       log("[Flow] Cannot flip checkbox when markdown contains a code block");
       return;
@@ -119,22 +100,12 @@ class DescriptionSection extends StatelessWidget {
     }
   }
 
-  void showEditModal(BuildContext context) async {
-    final String? result = await context.push<String?>(
-      "/utils/editmd",
-      extra: EditMarkdownPageProps(
-        initialValue: controller.text,
-        maxLength: Transaction.maxDescriptionLength,
-      ),
-    );
-
-    if (result == null) return;
-
-    controller.text = result;
-  }
-
   void onTapLink(
-      BuildContext context, String text, String? href, String title) {
+    BuildContext context,
+    String text,
+    String? href,
+    String title,
+  ) {
     log("[Flow] Tapped link: $text, $href, $title");
 
     if (href == null) {